pyproject.toml
[tool.poetry]
name = "scraper-agent"
version = "0.1.0"
description = "Agent AI to help with Web Scraping tasks"
authors = ["Khalen Fredieu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
chronological = "^0.1.1"
agent-protocol = "^1.0.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
typer = "^0.15.1"
beautifulsoup4 = "^4.12.3"
azure-identity = "^1.19.0"
msgraph-sdk = "^1.18.0"
pymongo = "^4.11"
transformers = {extras = ["torch"], version = "^4.48.2"}
torch = "^2.6.0"
scikit-learn = "^1.6.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
